move code from info module into a single file
