--- telemetry/src/info/mod.rs
@@ -1,1 +1,0 @@
-pub mod os;
